refactored to TypeScript
* [x] resolved https://github.com/0xadada/random-emoji/issues/2

refactored package to typescript

PR comments

thanks Katryn

simpler type
