+++ .gitignore
@@ -15,3 +15,4 @@ yarn-error.log
 
 # built assets
 /dist
+
