+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "outDir": "dist",
+    "strict": true
+  },
+  "exclude": ["node_modules", "test.js"]
+}
